curl is broken, at least on Synology DSM.
Somehow Synology curl refuses to work. It's very likely due to how the app was built. Synology is known to do "odd things".

The PR add a fallback to use wget.

To be honest, I'm not sure I like this solution. My very first solution, I had added a "curl check'. If it failed to talk to github, the script would fall back to wget. I actually prefer this solution as it might cover other NAS or Linux distributions.

Let me know.